Reorganised headers to fix compiler errors on Pi
